# src/main/resources/OpenApiDefinitions/pathfinding_streets-1.0.0-swagger.yaml
openapi: "3.1.0"
info:
  title: "osm_pathifinding API"
  description: "osm_pathifinding API"
  version: "1.0.0"
servers:
  - url: "https://localhost:8081"
tags:
  - name: "Streets and Addresses"
paths:
  /streets/all:
    get:
      summary: "All streets from database."
      description: ""
      operationId: "getStreetsFromDatabase"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/StreetDTO"
  /streets/:
    get:
      summary: "Get street by name from database."
      description: "GET streets?name={name}&number={number}"
      operationId: "getStreetByNameAndNumber"
      parameters:
        - name: "name"
          in: "query"
          required: true
          schema:
            type: "string"
        - name: "houseNumber"
          in: "query"
          required: false
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/StreetDTO"
  /streets/{id}:
    get:
      summary: "GET streets/{id}"
      operationId: "getStreetById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreetDTO"
  /streets/add:
    post:
      summary: "Add street."
      description: "Add street"
      operationId: "addStreet"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreetDTO"
        required: true
      responses:
        "200":
          description: "Street successfully added."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreetDTO"
  /streets/remove:
    delete:
      summary: "Remove street."
      description: "Remove street"
      operationId: "removeStreet"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreetDTO"
        required: true
      responses:
        "200":
          description: "Street successfully removed."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreetDTO"
components:
  schemas:
    NodeDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        longitude:
          type: "number"
          format: "double"
        latitude:
          type: "number"
          format: "double"
        streets:
          type: "array"
          items:
            $ref: "#/components/schemas/StreetDTO"
      required:
        - "id"
        - "longitude"
        - "latitude"
    StreetDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        isBuilding:
          type: "boolean"
        parent:
          $ref: "#/components/schemas/StreetDTO"
        children:
          type: "array"
          items:
            $ref: "#/components/schemas/StreetDTO"
        street:
          type: "string"
        housenumber:
          type: "string"
        nodes:
          type: "array"
          items:
            $ref: "#/components/schemas/NodeDTO"